Question7 - UI window
Added new UI window with the Jump button. Ran out of time to make the button move around. Seems like it starts at max pos x and can then be moved to the left with a time variable that I wasn't able to find. the Y position seems to be just radomized within the UI window. OnClick would just reset the variables.